backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart
celery==5.3.6
redis==5.0.1
qdrant-client
sentence-transformers==2.2.2
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
python-docx==1.1.0
pypdf==3.17.1
beautifulsoup4==4.12.2
pytesseract==0.3.10
Pillow==10.1.0
python-magic-bin==0.4.14
pytest==7.4.3
datasketch==1.6.4
